# Introduce Muon optimizer to PyTorch