--- a/benchmarks/muon_examples/train.py
+++ b/benchmarks/muon_examples/train.py
@@ -50,7 +50,6 @@ class MoonDataset(Dataset):
 
 # This code snippet is a modified version adapted from the following GitHub repository:
 # https://github.com/KellerJordan/Muon/blob/master/muon.py
-# @torch.compile
 def zeropower_via_newtonschulz5(G, steps):
     """
     Newton-Schulz iteration to compute the zeroth power / orthogonalization of G. We opt to use a