--- a/torch/optim/__init__.py
+++ b/torch/optim/__init__.py
@@ -26,6 +26,7 @@ from torch.optim.sparse_adam import SparseAdam as SparseAdam
 
 
 Adafactor.__module__ = "torch.optim"
+Muon.__module__ = "torch.optim"
 
 
 del adadelta  # type: ignore[name-defined] # noqa: F821