A single-device version of Muon. Algorithm refers to the [Moonshot](https://github.com/MoonshotAI/Moonlight) implementation.

**Usage**
This implementation requires users to pass in `named_parameters` of the model - a list of `(name, param)` tuples. Users should also specify the fully-qualified names (FQNs) of the parameters to be optimized by Muon. Parameters not included in the FQN list will fall back to AdamW optimization. If no FQN list is provided, Muon will by default optimize 2D parameters, which may not be the expected behavior. A warning will be issued in this case.

```python
    model = MyModelForCausalLM
    muon_param_fqns = [
        "model.layers.0.self_attn.q_proj.weight", 
        "model.layers.0.mlp.up_proj.weight"
    ]
    muon = Muon(
        model.named_parameters(),
        lr=lr,
        wd=wd,
        muon_param_fqns=muon_param_fqns,
        adamw_betas=adamw_betas,
        adamw_eps=adamw_eps,
    )
```

**Additional usage**
Users are also able to pass in self-defined `msign` function for orthogonalization, and learning rate adjustment function. Interface defined below:

```python
AdjustLrFn: TypeAlias = Callable[[float, torch.Size], float]
MsignFn: TypeAlias = Callable[[Tensor, BaseMsignFnConfig], Tensor]
```
By default, we use 5-step Newton-Schulz, with coefficients proposed by [Keller](https://kellerjordan.github.io/posts/muon/). We use LR adjustment proposed by [Moonshot](https://github.com/MoonshotAI/Moonlight/blob/master/Moonlight.pdf), which grafts learning rate from AdamW.

**Testing**

1. Unit tests: the newly introduced Muon is covered in `test/test_optim.py`. We updated the test cases to pass named parameters to the optimizer under test. Additionally, we introduced a new test case to verify that when the user provides an empty FQN list, Muon correctly falls back to AdamW behavior.
2. End-to-end test: we added a training script that pre-trains a QWEN-like model on `openwebtext-100k` dataset. We trained for one epoch and the resulting loss curve is compared against the Moonshot implementation to confirm behavioral consistency.
<img width="1102" height="472" alt="Screenshot 2025-07-29 at 1 04 12 AM" src="https://github.com/user-attachments/assets/ceab0733-497d-4070-8032-02ae7995c64c" />

**Performance**
Training for one epoch of openwebtext-100k on eight H100 GPUs with DDP:

- adamw_ddp finishes in 13.12 min
- pytorch_muon_ddp finishes in 13.45 min

Muon runs ~20s slower compared to AdamW. Assuming no other changes, Muon is *2.5%* slower than AdamW.

AdamW: Optimizer.step() takes ~13.5 ms, step time ~930 ms
<img width="726" height="590" alt="Screenshot 2025-07-29 at 1 56 14 AM" src="https://github.com/user-attachments/assets/ebcd7e1c-d129-4b20-9396-39f568edf03d" />

Muon: Optimizer.step() takes ~54 ms, step time ~960 ms
<img width="751" height="597" alt="Screenshot 2025-07-29 at 2 02 20 AM" src="https://github.com/user-attachments/assets/72f5b904-ebd5-4502-a6ff-d3e9e5a6da81" />

**Next Steps**

1. Add `MuP`
2. Open-source optimized triton kernel for symmetric matmul. A preliminary benchmark found 1.23x - 1.48x speedup on small - large (n = 256 -> 16384) matrices.
3. Open-source unsharded Muon co-designed with FSDP2.
 
cc: @toothacher17, @vinaysrao, @jcui2, @haocizhang

cc @voznesenskym @penguinwu @EikanWang @jgong5 @Guobing-Chen @XiaobingSuper @zhuhaozhe @blzheng @wenzhe-nrv @jiayisunx @ipiszy @chenyang78 @kadeng @muchulee8 @amjames @chauhang @aakhundov @coconutruben